fix duplicates
